Inline the ExtensionPageInjector
